extract out common code to reduce duplication
